Created separate leaf render object widget for chart legends
