+++ lib/geekyants_flutter_charts.dart
@@ -3,3 +3,4 @@ library geekyants_flutter_charts;
 export 'src/bar_chart/bar_chart.dart';
 export 'src/bar_chart/bar_chart_axis.dart';
 export 'src/bar_chart/bar_chart_title.dart';
+export 'src/bar_chart/bar_chart_legends.dart';
