+++ lib/src/bar_chart/bar_chart_axis.dart
@@ -7,12 +7,6 @@ import 'dart:math' as math;
 /// It takes several parameters that define the size and layout of the chart, as
 /// well as the number of axis points and the bar width.
 class BarChartAxis extends LeafRenderObjectWidget {
-  final String legendTextXAxis;
-  final String legendTextYAxis;
-  final TextStyle legendTextXAxisStyle;
-  final TextStyle legendTextYAxisStyle;
-  final Color legendXAxisColor;
-  final Color legendYAxisColor;
   final double xAxisRulerHeight;
   final double yAxisRulerHeight;
   final double xAxisRulerThickness;
@@ -53,14 +47,6 @@ class BarChartAxis extends LeafRenderObjectWidget {
 
   const BarChartAxis({
     Key? key,
-    this.legendTextXAxis = "X-Axis",
-    this.legendTextYAxis = "Y-Axis Value",
-    this.legendTextXAxisStyle =
-        const TextStyle(color: Colors.black, fontSize: 10),
-    this.legendTextYAxisStyle =
-        const TextStyle(color: Colors.black, fontSize: 10),
-    this.legendXAxisColor = Colors.blue,
-    this.legendYAxisColor = Colors.black,
     this.xAxisRulerHeight = 10,
     this.yAxisRulerHeight = 10,
     this.xAxisRulerThickness = 0.2,
@@ -101,12 +87,6 @@ class BarChartAxis extends LeafRenderObjectWidget {
   @override
   RenderObject createRenderObject(BuildContext context) {
     return RenderBarChartAxis(
-      legendTextXAxis: legendTextXAxis,
-      legendTextYAxis: legendTextYAxis,
-      legendTextXAxisStyle: legendTextXAxisStyle,
-      legendTextYAxisStyle: legendTextYAxisStyle,
-      legendXAxisColor: legendXAxisColor,
-      legendYAxisColor: legendYAxisColor,
       xAxisRulerHeight: xAxisRulerHeight,
       yAxisRulerHeight: yAxisRulerHeight,
       xAxisRulerThickness: xAxisRulerThickness,
@@ -146,12 +126,6 @@ class BarChartAxis extends LeafRenderObjectWidget {
   void updateRenderObject(
       BuildContext context, covariant RenderBarChartAxis renderObject) {
     renderObject
-      ..legendTextXAxis = legendTextXAxis
-      ..legendTextYAxis = legendTextYAxis
-      ..legendTextXAxisStyle = legendTextXAxisStyle
-      ..legendTextYAxisStyle = legendTextYAxisStyle
-      ..legendXAxisColor = legendXAxisColor
-      ..legendYAxisColor = legendYAxisColor
       ..xAxisRulerHeight = xAxisRulerHeight
       ..yAxisRulerHeight = yAxisRulerHeight
       ..xAxisRulerThickness = xAxisRulerThickness
@@ -197,12 +171,6 @@ class RenderBarChartAxis extends RenderBox {
   ///
   /// The [context] parameter provides the painting context, and the [offset]
   /// parameter specifies the offset at which to paint the axes.
-  String renderLegendTextXAxis;
-  String renderLegendTextYAxis;
-  TextStyle renderLegendTextXAxisStyle;
-  TextStyle renderLegendTextYAxisStyle;
-  Color renderLegendXAxisColor;
-  Color renderLegendYAxisColor;
   double renderXAxisRulerHeight;
   double renderYAxisRulerHeight;
   TextStyle renderXAxisLabelTextStyle;
@@ -236,12 +204,6 @@ class RenderBarChartAxis extends RenderBox {
   double renderYAxisLabelOffset;
   List<double> renderYAxisData;
   RenderBarChartAxis({
-    required String legendTextXAxis,
-    required String legendTextYAxis,
-    required TextStyle legendTextXAxisStyle,
-    required TextStyle legendTextYAxisStyle,
-    required Color legendXAxisColor,
-    required Color legendYAxisColor,
     required double xAxisRulerHeight,
     required double yAxisRulerHeight,
     required double xAxisRulerThickness,
@@ -274,13 +236,7 @@ class RenderBarChartAxis extends RenderBox {
     required double xAxisSteps,
     required double yAxisSteps,
     required List<double> yAxisData,
-  })  : renderLegendTextXAxis = legendTextXAxis,
-        renderLegendTextYAxis = legendTextYAxis,
-        renderLegendTextXAxisStyle = legendTextXAxisStyle,
-        renderLegendTextYAxisStyle = legendTextYAxisStyle,
-        renderLegendXAxisColor = legendXAxisColor,
-        renderLegendYAxisColor = legendYAxisColor,
-        renderXAxisRulerHeight = xAxisRulerHeight,
+  })  : renderXAxisRulerHeight = xAxisRulerHeight,
         renderYAxisRulerHeight = yAxisRulerHeight,
         renderXAxisRulerThickness = xAxisRulerThickness,
         renderYAxisRulerThickness = yAxisRulerThickness,
@@ -316,54 +272,6 @@ class RenderBarChartAxis extends RenderBox {
   double _thicknessOfYAxis = 0;
   Size sizeOfXAxisLabel = Size.zero;
   double extendXAxisStart = 0;
-
-  String get legendTextXAxis => renderLegendTextXAxis;
-  set legendTextXAxis(String value) {
-    if (renderLegendTextXAxis != value) {
-      renderLegendTextXAxis = value;
-      markNeedsLayout();
-    }
-  }
-
-  String get legendTextYAxis => renderLegendTextYAxis;
-  set legendTextYAxis(String value) {
-    if (renderLegendTextYAxis != value) {
-      renderLegendTextYAxis = value;
-      markNeedsLayout();
-    }
-  }
-
-  TextStyle get legendTextXAxisStyle => renderLegendTextXAxisStyle;
-  set legendTextXAxisStyle(TextStyle value) {
-    if (renderLegendTextXAxisStyle != value) {
-      renderLegendTextXAxisStyle = value;
-      markNeedsLayout();
-    }
-  }
-
-  TextStyle get legendTextYAxisStyle => renderLegendTextYAxisStyle;
-  set legendTextYAxisStyle(TextStyle value) {
-    if (renderLegendTextYAxisStyle != value) {
-      renderLegendTextYAxisStyle = value;
-      markNeedsLayout();
-    }
-  }
-
-  Color get legendXAxisColor => renderLegendXAxisColor;
-  set legendXAxisColor(Color value) {
-    if (renderLegendXAxisColor != value) {
-      renderLegendXAxisColor = value;
-      markNeedsPaint();
-    }
-  }
-
-  Color get legendYAxisColor => renderLegendYAxisColor;
-  set legendYAxisColor(Color value) {
-    if (renderLegendYAxisColor != value) {
-      renderLegendYAxisColor = value;
-      markNeedsPaint();
-    }
-  }
 
   double get xAxisRulerHeight => renderXAxisRulerHeight;
   set xAxisRulerHeight(double value) {
@@ -624,12 +532,11 @@ class RenderBarChartAxis extends RenderBox {
   @override
   void paint(PaintingContext context, Offset offset) {
     final Canvas canvas = context.canvas;
-    final double graphWidth = size.width;
+    final double graphWidth = size.width - offset.dx;
     final double graphHeight = size.height - offset.dy;
     double yAxisMainLabelOffset = 0;
     List<AxesLabel> xAxesLabel = [];
     List<AxesLabel> yAxesLabel = [];
-    double legendWidth;
     calculateLabelValues(
         xAxisStartPoint, xAxisEndPoint, graphWidth, xAxisSteps, xAxesLabel);
     sizeOfXAxisLabel = getLabelSize(
@@ -671,58 +578,9 @@ class RenderBarChartAxis extends RenderBox {
       value: yAxesLabel.last.value.toString(),
     );
 
-    /// Legends for the chart
-    final textPainter = TextPainter(textDirection: TextDirection.ltr);
-    final circlePaint = Paint();
-
-    final List<String> legendTexts = [legendTextXAxis, legendTextYAxis];
-    final List<TextStyle> legendTextStyles = [
-      legendTextXAxisStyle,
-      legendTextYAxisStyle
-    ];
-    double maxWidth = 0;
-
-    for (int i = 0; i < legendTexts.length; i++) {
-      textPainter.text =
-          TextSpan(text: legendTexts[i], style: legendTextStyles[i]);
-      textPainter.layout();
-
-      double width = textPainter.width;
-      if (width > maxWidth) {
-        maxWidth = width;
-      }
-    }
-
-    paintLegends(
-        canvas,
-        legendTextXAxis,
-        legendTextXAxisStyle,
-        Offset(size.width - maxWidth, size.height / 2),
-        legendXAxisColor,
-        textPainter,
-        circlePaint);
-
-    paintLegends(
-        canvas,
-        legendTextYAxis,
-        legendTextYAxisStyle,
-        Offset(
-            size.width - maxWidth, size.height / 2 + textPainter.height + 10),
-        legendYAxisColor,
-        textPainter,
-        circlePaint);
-
-    // Formula to find diameter of the circle
-    const circleWidth = 2 * math.pi * 3;
-
-    final legendWidthX = maxWidth + circleWidth;
-    final legendWidthY = maxWidth + circleWidth;
-
-    legendWidth = legendWidthX > legendWidthY ? legendWidthX : legendWidthY;
-
     paintYAxisLabels(
         graphHeight,
-        graphWidth - legendWidth,
+        graphWidth,
         yAxisMainLabelOffset,
         yAxisLabelOffset,
         yAxisRulerOffset,
@@ -737,7 +595,7 @@ class RenderBarChartAxis extends RenderBox {
         showYAxisGridRuler);
 
     paintXAxisLabels(
-        graphWidth - legendWidth,
+        graphWidth,
         xAxesLabel,
         canvas,
         yStartLabelSize,
@@ -748,26 +606,6 @@ class RenderBarChartAxis extends RenderBox {
         offset,
         showXAxisGridRuler,
         yAxesLabel);
-  }
-
-  void paintLegends(
-      Canvas canvas,
-      String text,
-      TextStyle textStyle,
-      Offset textOffset,
-      Color circleColor,
-      TextPainter textPainter,
-      Paint circlePaint) {
-    textPainter.text = TextSpan(text: text, style: textStyle);
-    textPainter.layout();
-    textPainter.paint(canvas, textOffset);
-
-    final circleOffset = Offset(
-      textOffset.dx - textStyle.fontSize! / 2,
-      textOffset.dy + textPainter.height / 2,
-    );
-    circlePaint.color = circleColor;
-    canvas.drawCircle(circleOffset, 3, circlePaint);
   }
 
   void paintXAxisLabels(
